+++ src/Apps/Order2/Routes/Checkout/Components/__tests__/Order2HelpLinks.jest.tsx
@@ -1,5 +1,5 @@
 import { fireEvent, screen } from "@testing-library/dom"
-import { Order2HelpLinks } from "Apps/Order2/Routes/Checkout/Components/Order2HelpLinks"
+import { Order2HelpLinks } from "Apps/Order2/Components/Order2HelpLinks"
 import { setupTestWrapperTL } from "DevTools/setupTestWrapperTL"
 import type { Order2DetailsPage_Test_Query } from "__generated__/Order2DetailsPage_Test_Query.graphql"
 import { useTracking } from "react-tracking"
